requirements/dev.txt
-r common.txt

bumpversion==0.5.3
django-debug-toolbar==1.6
django-nose==1.4.4
factory-boy==2.8.1
nose==1.3.7
pylama==7.2.3
